refactor(common): split market API into market and countervalues RTK Query APIs
- Split market api layer into market.ts (MARKET_API_URL) and countervalues.ts (LEDGER_COUNTERVALUES_API)
- Split state-manager into marketApi (coins list, counter currencies, chart data) and countervaluesApi (performers)
- useMarketDataProvider now uses RTK Query hooks for chart data and supported currencies
- Register countervaluesApi in desktop and mobile RTK Query reducers

LIVE-25622
