+++ .changeset/spotty-seas-admire.md
@@ -1,0 +1,7 @@
+---
+"ledger-live-desktop": minor
+"live-mobile": minor
+"@ledgerhq/live-common": minor
+---
+
+Split market API into market and countervalues RTK Query APIs
