+++ apps/ledger-live-desktop/src/renderer/reducers/rtkQueryApi.ts
@@ -1,7 +1,7 @@
 import type { Middleware, Reducer, Tuple } from "@reduxjs/toolkit";
 import { ofacGeoBlockApi } from "@ledgerhq/live-common/api/ofacGeoBlockApi";
 import { assetsDataApi } from "@ledgerhq/live-common/dada-client/state-manager/api";
-import { marketApi } from "@ledgerhq/live-common/market/state-manager/api";
+import { countervaluesApi, marketApi } from "@ledgerhq/live-common/market/state-manager/api";
 import { cryptoAssetsApi } from "@ledgerhq/cryptoassets/cal-client/state-manager/api";
 import { pushDevicesApi } from "@ledgerhq/client-ids/api";
 import { cmcApi } from "@ledgerhq/live-common/cmc-client/state-manager/api";
@@ -11,6 +11,7 @@ const APIs = {
   [assetsDataApi.reducerPath]: assetsDataApi,
   [cmcApi.reducerPath]: cmcApi,
   [cryptoAssetsApi.reducerPath]: cryptoAssetsApi,
+  [countervaluesApi.reducerPath]: countervaluesApi,
   [marketApi.reducerPath]: marketApi,
   [ofacGeoBlockApi.reducerPath]: ofacGeoBlockApi,
   [pushDevicesApi.reducerPath]: pushDevicesApi,
