+++ libs/ledger-live-common/.unimportedrc.json
@@ -466,6 +466,10 @@
     "src/market/utils/types.ts",
     "src/market/state-manager/api.ts",
     "src/market/state-manager/types.ts",
+    "src/market/api/countervalues.ts",
+    "src/market/api/market.ts",
+    "src/market/state-manager/countervaluesApi.ts",
+    "src/market/state-manager/marketApi.ts",
     "src/modularDrawer/enums.ts",
     "src/modularDrawer/hooks/index.ts",
     "src/modularDrawer/hooks/modules/useLeftApyModule.tsx",
