+++ libs/ledger-live-common/src/market/hooks/useMarketDataProvider.ts
@@ -1,11 +1,10 @@
 import { UseQueryResult, useQueries, useQuery } from "@tanstack/react-query";
+import { fetchCurrency, fetchList } from "../api";
 import {
-  fetchCurrency,
-  fetchCurrencyChartData,
-  fetchList,
-  getSupportedCoinsList,
-  supportedCounterCurrencies,
-} from "../api";
+  useGetSupportedCoinsListQuery,
+  useGetSupportedCounterCurrenciesQuery,
+  useGetCurrencyChartDataQuery,
+} from "../state-manager/api";
 import { currencyFormatter, format } from "../utils/currencyFormatter";
 import { QUERY_KEY } from "../utils/queryKeys";
 import { REFETCH_TIME_ONE_MINUTE, BASIC_REFETCH, ONE_DAY } from "../utils/timers";
@@ -30,11 +29,21 @@ export function useMarketDataProvider() {
 }
 
 export const useCurrencyChartData = ({ id, counterCurrency, range }: MarketCurrencyRequestParams) =>
-  useQuery({
-    queryKey: [QUERY_KEY.CurrencyChartData, id, counterCurrency, range],
-    queryFn: () => fetchCurrencyChartData({ counterCurrency, range, id }),
-    refetchInterval: REFETCH_TIME_ONE_MINUTE * BASIC_REFETCH,
-    staleTime: REFETCH_TIME_ONE_MINUTE * BASIC_REFETCH,
+  useGetCurrencyChartDataQuery(
+    { id, counterCurrency, range },
+    {
+      pollingInterval: REFETCH_TIME_ONE_MINUTE * BASIC_REFETCH,
+    },
+  );
+
+export const useSupportedCounterCurrencies = () =>
+  useGetSupportedCounterCurrenciesQuery(undefined, {
+    pollingInterval: ONE_DAY,
+  });
+
+export const useSupportedCurrencies = () =>
+  useGetSupportedCoinsListQuery(undefined, {
+    pollingInterval: ONE_DAY,
   });
 
 export const useCurrencyData = ({ id, counterCurrency }: MarketCurrencyRequestParams) =>
@@ -44,22 +53,6 @@ export const useCurrencyData = ({ id, counterCurrency }: MarketCurrencyRequestPa
     refetchInterval: REFETCH_TIME_ONE_MINUTE * BASIC_REFETCH,
     staleTime: REFETCH_TIME_ONE_MINUTE * BASIC_REFETCH,
     select: data => format(data),
-  });
-
-export const useSupportedCounterCurrencies = () =>
-  useQuery({
-    queryKey: [QUERY_KEY.SupportedCounterCurrencies],
-    queryFn: () => supportedCounterCurrencies(),
-    refetchOnWindowFocus: true,
-    staleTime: ONE_DAY,
-  });
-
-export const useSupportedCurrencies = () =>
-  useQuery({
-    queryKey: [QUERY_KEY.SupportedCurrencies],
-    queryFn: () => getSupportedCoinsList(),
-    refetchOnWindowFocus: true,
-    staleTime: ONE_DAY,
   });
 
 export function useMarketData(props: MarketListRequestParams): MarketListRequestResult {
