+++ libs/ledger-live-common/src/market/state-manager/types.ts
@@ -2,6 +2,9 @@ import { PortfolioRange } from "@ledgerhq/types-live";
 
 export enum MarketDataTags {
   Performers = "Performers",
+  Coins = "Coins",
+  CounterCurrencies = "CounterCurrencies",
+  ChartData = "ChartData",
 }
 
 export interface MarketPerformersQueryParams {
